Sun, Mar 17, 2024, 1:02 AM +01:00

+++ credit/credit.c
@@ -16,7 +16,8 @@ int main(void)
 
     for (int i = len - 2; i >= 0; i-=2)
     {
-        sum_mul += strnum[i] * 2;
+
+        sum_mul += (int)strnum[i] * 2;
     }
 
     for (int i = len - 1; i >= 0; i-=2)
